Migration fix

+++ database/migrations/2020_03_20_183937_add_beneficiary_id_to_solicitations_table.php
@@ -17,7 +17,7 @@ class AddBeneficiaryIdToSolicitationsTable extends Migration
             $table->string('payment_type');
             $table->string('payment_status');
             $table->decimal('payment_value');
-            $table->integer('beneficiary_id')->index()->unsigned();
+            $table->bigInteger('beneficiary_id')->index()->unsigned();
             $table->foreign('beneficiary_id')->references('id')->on('beneficiaries');
         });
     }
